Major code cleanup, formatting and styling
git-svn-id: https://svn.apache.org/repos/asf/turbine/fulcrum/trunk/intake@670328 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fulcrum/intake/model/ShortField.java
@@ -139,7 +139,7 @@ public class ShortField
             
             if (value == null)
             {
-                setTestValue((Short)getEmptyValue());
+                setTestValue(getEmptyValue());
             }
             else
             {
